add event emitter and listener, update groups controller and update all services. wip calculate balance

+++ src/modules/expense-sharing/services/expense-contributor.service.ts
@@ -5,12 +5,15 @@ import { ExpenseContributor } from '../entities/expense-contributor.entity';
 import { UpdateExpenseContributorDto } from '../dto/requests/update-expense-contributor-dto';
 import { CreateExpenseContributorDto } from '../dto/requests/create-expense-contributor-dto';
 import { ExpenseGroupMemberService } from './expense-group-member.service';
+import { EventEmitter2 } from '@nestjs/event-emitter';
+import { EXPENSE_CHANGED_EVENT } from '../events/expense.events';
 
 @Injectable()
 export class ExpenseContributorService {
     constructor(
         @InjectRepository(ExpenseContributor) private readonly contributorRepository: Repository<ExpenseContributor>,
         private readonly memberService: ExpenseGroupMemberService,
+        private readonly eventEmitter: EventEmitter2,
     ) {}
 
     async findAll(expenseId: string){
@@ -39,6 +42,10 @@ export class ExpenseContributorService {
 
         await this.contributorRepository.save(contributor);
 
+        this.eventEmitter.emit(EXPENSE_CHANGED_EVENT, {
+            expenseId,
+        });
+
         return contributor;
     }
 
@@ -51,6 +58,10 @@ export class ExpenseContributorService {
             }
         );
 
+        this.eventEmitter.emit(EXPENSE_CHANGED_EVENT, {
+            expenseId,
+        });
+
         const updatedContributor = await this.contributorRepository.findOneBy({ id: contributorId });
         if(!updatedContributor) throw new HttpException('Updated contributor not found', HttpStatus.NOT_FOUND)
 
@@ -58,6 +69,15 @@ export class ExpenseContributorService {
     }
 
     async delete(id: string) {
-        return await this.contributorRepository.delete(id);
+        const contributor = await this.contributorRepository.findOneBy({ id })
+        if(!contributor) throw new HttpException('Contributor not found', HttpStatus.NOT_FOUND)
+
+        await this.contributorRepository.delete(id);
+
+        this.eventEmitter.emit(EXPENSE_CHANGED_EVENT, {
+            expenseId: contributor.expenseId,
+        });
+
+        return { success: true };
     }
 }
